+++ apps/backend/src/libs/modules/server-application/server-application.ts
@@ -4,6 +4,7 @@ import { logger } from "~/libs/modules/logger/logger.js";
 import { activityLogController } from "~/modules/activity-logs/activity-logs.js";
 import { authController } from "~/modules/auth/auth.js";
 import { groupController } from "~/modules/groups/groups.js";
+import { permissionController } from "~/modules/permissions/permissions.js";
 import { projectGroupController } from "~/modules/project-groups/project-groups.js";
 import { projectController } from "~/modules/projects/projects.js";
 import { userController, userService } from "~/modules/users/users.js";
@@ -18,6 +19,7 @@ const apiV1 = new BaseServerApplicationApi(
 	config,
 	...activityLogController.routes,
 	...authController.routes,
+	...permissionController.routes,
 	...projectGroupController.routes,
 	...projectController.routes,
 	...userController.routes,