+++ apps/backend/src/modules/groups/group.repository.ts
@@ -41,12 +41,19 @@ class GroupRepository implements Repository {
 		return GroupEntity.initialize(group);
 	}
 
-	public delete(): ReturnType<Repository["delete"]> {
-		return Promise.resolve(true);
+	public async delete(id: number): Promise<boolean> {
+		const deletedRowsCount = await this.groupModel
+			.query()
+			.deleteById(id)
+			.execute();
+
+		return Boolean(deletedRowsCount);
 	}
 
-	public find(): ReturnType<Repository["find"]> {
-		return Promise.resolve(null);
+	public async find(id: number): Promise<GroupModel | null> {
+		const group = await this.groupModel.query().findById(id).execute();
+
+		return group ?? null;
 	}
 
 	public async findAll({