+++ apps/backend/src/modules/users/user.entity.ts
@@ -1,26 +1,28 @@
 import { type Entity } from "~/libs/types/types.js";
+import { type GroupModel } from "~/modules/groups/group.model.js";
+import { type PermissionModel } from "~/modules/permissions/permission.model.js";
 
 import { type UserAuthResponseDto } from "./libs/types/types.js";
 
 class UserEntity implements Entity {
 	private createdAt: null | string;
-
 	private deletedAt: null | string;
-
 	private email: string;
-
+	private groups: Array<
+		{
+			permissions: Array<Pick<PermissionModel, "id" | "key" | "name">>;
+		} & Pick<GroupModel, "id" | "name">
+	>;
 	private id: null | number;
-
 	private name: string;
-
 	private passwordHash: string;
-
 	private passwordSalt: string;
 
 	private constructor({
 		createdAt,
 		deletedAt,
 		email,
+		groups = [],
 		id,
 		name,
 		passwordHash,
@@ -29,6 +31,11 @@ class UserEntity implements Entity {
 		createdAt: null | string;
 		deletedAt: null | string;
 		email: string;
+		groups?: Array<
+			{
+				permissions: Array<Pick<PermissionModel, "id" | "key" | "name">>;
+			} & Pick<GroupModel, "id" | "name">
+		>;
 		id: null | number;
 		name: string;
 		passwordHash: string;
@@ -36,6 +43,7 @@ class UserEntity implements Entity {
 	}) {
 		this.id = id;
 		this.email = email;
+		this.groups = groups;
 		this.name = name;
 		this.passwordHash = passwordHash;
 		this.passwordSalt = passwordSalt;
@@ -47,6 +55,7 @@ class UserEntity implements Entity {
 		createdAt,
 		deletedAt,
 		email,
+		groups = [],
 		id,
 		name,
 		passwordHash,
@@ -55,6 +64,11 @@ class UserEntity implements Entity {
 		createdAt: string;
 		deletedAt: null | string;
 		email: string;
+		groups?: Array<
+			{
+				permissions: Array<Pick<PermissionModel, "id" | "key" | "name">>;
+			} & Pick<GroupModel, "id" | "name">
+		>;
 		id: number;
 		name: string;
 		passwordHash: string;
@@ -64,6 +78,7 @@ class UserEntity implements Entity {
 			createdAt,
 			deletedAt,
 			email,
+			groups,
 			id,
 			name,
 			passwordHash,
@@ -78,6 +93,11 @@ class UserEntity implements Entity {
 		passwordSalt,
 	}: {
 		email: string;
+		groups: Array<
+			{
+				permissions: Array<Pick<PermissionModel, "id" | "key" | "name">>;
+			} & Pick<GroupModel, "id" | "name">
+		>;
 		name: string;
 		passwordHash: string;
 		passwordSalt: string;
@@ -86,6 +106,7 @@ class UserEntity implements Entity {
 			createdAt: null,
 			deletedAt: null,
 			email,
+			groups: [],
 			id: null,
 			name,
 			passwordHash,
@@ -95,12 +116,14 @@ class UserEntity implements Entity {
 
 	public toNewObject(): {
 		email: string;
+		groups: Pick<GroupModel, "id" | "name">[];
 		name: string;
 		passwordHash: string;
 		passwordSalt: string;
 	} {
 		return {
 			email: this.email,
+			groups: this.groups,
 			name: this.name,
 			passwordHash: this.passwordHash,
 			passwordSalt: this.passwordSalt,
@@ -111,6 +134,15 @@ class UserEntity implements Entity {
 		return {
 			createdAt: this.createdAt as string,
 			email: this.email,
+			groups: this.groups.map((group) => ({
+				id: group.id,
+				name: group.name,
+				permissions: group.permissions.map((permission) => ({
+					id: permission.id,
+					key: permission.key,
+					name: permission.name,
+				})),
+			})),
 			id: this.id as number,
 			name: this.name,
 		};