+++ apps/backend/src/modules/users/user.repository.ts
@@ -3,10 +3,9 @@ import {
 	type PaginationResponseDto,
 	type Repository,
 } from "~/libs/types/types.js";
+import { type UserPatchRequestDto } from "~/modules/users/libs/types/types.js";
 import { UserEntity } from "~/modules/users/user.entity.js";
 import { type UserModel } from "~/modules/users/user.model.js";
-
-import { type UserPatchRequestDto } from "./libs/types/types.js";
 
 class UserRepository implements Repository {
 	private userModel: typeof UserModel;
@@ -47,7 +46,13 @@ class UserRepository implements Repository {
 			.query()
 			.findById(id)
 			.whereNull("deletedAt")
-			.returning("*")
+			.withGraphFetched("groups.permissions")
+			.modifyGraph("groups", (builder) => {
+				builder.select("user_groups.id", "name");
+			})
+			.modifyGraph("groups.permissions", (builder) => {
+				builder.select("permissions.id", "name", "key");
+			})
 			.execute();
 
 		return user ? UserEntity.initialize(user) : null;
@@ -59,6 +64,7 @@ class UserRepository implements Repository {
 	}: PaginationQueryParameters): Promise<PaginationResponseDto<UserEntity>> {
 		const { results, total } = await this.userModel
 			.query()
+			.withGraphFetched("groups.permissions")
 			.whereNull("deletedAt")
 			.page(page, pageSize);
 
@@ -72,7 +78,16 @@ class UserRepository implements Repository {
 		email: string,
 		hasDeleted = false,
 	): Promise<null | UserEntity> {
-		const query = this.userModel.query().findOne({ email });
+		const query = this.userModel
+			.query()
+			.findOne({ email })
+			.withGraphFetched("groups.permissions")
+			.modifyGraph("user_groups.id", (builder) => {
+				builder.select("id", "name");
+			})
+			.modifyGraph("groups.permissions", (builder) => {
+				builder.select("permissions.id", "name", "key");
+			});
 
 		if (!hasDeleted) {
 			query.whereNull("deletedAt");