+++ apps/frontend/src/libs/components/popover/popover.tsx
@@ -1,10 +1,15 @@
-import { useHandleClickOutside, useRef } from "~/libs/hooks/hooks.js";
+import {
+	useHandleClickOutside,
+	usePopoverPosition,
+	useRef,
+} from "~/libs/hooks/hooks.js";
 
 import styles from "./styles.module.css";
 
 type Properties = {
 	children: React.ReactNode;
 	content: React.ReactNode;
+	hasFixedPositioning?: boolean;
 	isOpened: boolean;
 	onClose: () => void;
 };
@@ -12,18 +17,40 @@ type Properties = {
 const Popover = ({
 	children,
 	content,
+	hasFixedPositioning,
 	isOpened,
 	onClose,
 }: Properties): JSX.Element => {
 	const popoverReference = useRef<HTMLDivElement>(null);
+	const popoverTargetReference = useRef<HTMLDivElement>(null);
+
+	const position = usePopoverPosition(
+		popoverTargetReference,
+		isOpened,
+		hasFixedPositioning,
+	);
 
 	useHandleClickOutside(popoverReference, onClose);
 
 	return (
 		<div className={styles["popover-wrapper"]} ref={popoverReference}>
 			{children}
+
 			{isOpened && (
-				<div className={styles["popover-content-wrapper"]}>{content}</div>
+				<div
+					className={styles["popover-content-wrapper"]}
+					ref={popoverTargetReference}
+				>
+					<div
+						className={styles["popover-content"]}
+						style={{
+							left: position.left,
+							top: position.top,
+						}}
+					>
+						{content}
+					</div>
+				</div>
 			)}
 		</div>
 	);