+++ apps/frontend/src/libs/enums/notification-message.enum.ts
@@ -1,6 +1,8 @@
 const NotificationMessage = {
-	PROFILE_UPDATE_SUCCESS: "Successfully updated profile information.",
-	PROJECT_CREATE_SUCCESS: "Project was successfully created",
+	GROUP_CREATE_SUCCESS: "Group was successfully created.",
+	GROUP_DELETE_SUCCESS: "Group was successfully deleted.",
+	PROFILE_UPDATE_SUCCESS: "Your profile has been successfully updated.",
+	PROJECT_CREATE_SUCCESS: "Project was successfully created.",
 	PROJECT_DELETE_SUCCESS: "Project was successfully deleted.",
 	PROJECT_UPDATE_SUCCESS: "Project was successfully updated.",
 	USER_DELETE_SUCCESS: "User deleted successfully.",