+++ apps/frontend/src/modules/groups/slices/group.slice.ts
@@ -1,25 +1,51 @@
 import { createSlice } from "@reduxjs/toolkit";
 
+import { ITEMS_CHANGED_COUNT } from "~/libs/components/table-pagination/libs/constants/constants.js";
 import { DataStatus } from "~/libs/enums/enums.js";
 import { type ValueOf } from "~/libs/types/types.js";
+import { type UserGetAllItemResponseDto } from "~/modules/users/users.js";
 
 import { type GroupGetAllItemResponseDto } from "../libs/types/types.js";
-import { loadAll } from "./actions.js";
+import { create, deleteById, loadAll, loadUsers } from "./actions.js";
 
 type State = {
 	dataStatus: ValueOf<typeof DataStatus>;
+	groupCreateStatus: ValueOf<typeof DataStatus>;
+	groupDeleteStatus: ValueOf<typeof DataStatus>;
 	groups: GroupGetAllItemResponseDto[];
 	groupsTotalCount: number;
+	users: UserGetAllItemResponseDto[];
+	usersDataStatus: ValueOf<typeof DataStatus>;
+	usersTotalCount: number;
 };
 
 const initialState: State = {
 	dataStatus: DataStatus.IDLE,
+	groupCreateStatus: DataStatus.IDLE,
+	groupDeleteStatus: DataStatus.IDLE,
 	groups: [],
 	groupsTotalCount: 0,
+	users: [],
+	usersDataStatus: DataStatus.IDLE,
+	usersTotalCount: 0,
 };
 
 const { actions, name, reducer } = createSlice({
 	extraReducers(builder) {
+		builder.addCase(loadUsers.pending, (state) => {
+			state.usersDataStatus = DataStatus.PENDING;
+		});
+
+		builder.addCase(loadUsers.fulfilled, (state, action) => {
+			state.users = action.payload.items;
+			state.usersTotalCount = action.payload.totalItems;
+			state.usersDataStatus = DataStatus.FULFILLED;
+		});
+
+		builder.addCase(loadUsers.rejected, (state) => {
+			state.usersDataStatus = DataStatus.REJECTED;
+		});
+
 		builder.addCase(loadAll.pending, (state) => {
 			state.dataStatus = DataStatus.PENDING;
 		});
@@ -31,6 +57,31 @@ const { actions, name, reducer } = createSlice({
 		builder.addCase(loadAll.rejected, (state) => {
 			state.groups = [];
 			state.dataStatus = DataStatus.REJECTED;
+		});
+
+		builder.addCase(deleteById.pending, (state) => {
+			state.groupDeleteStatus = DataStatus.PENDING;
+		});
+		builder.addCase(deleteById.fulfilled, (state, action) => {
+			const { id } = action.meta.arg;
+			state.groups = state.groups.filter((group) => group.id !== id);
+			state.groupsTotalCount -= ITEMS_CHANGED_COUNT;
+			state.groupDeleteStatus = DataStatus.FULFILLED;
+		});
+		builder.addCase(deleteById.rejected, (state) => {
+			state.groupDeleteStatus = DataStatus.REJECTED;
+		});
+
+		builder.addCase(create.pending, (state) => {
+			state.groupCreateStatus = DataStatus.PENDING;
+		});
+		builder.addCase(create.fulfilled, (state, action) => {
+			state.groups = [action.payload, ...state.groups];
+			state.groupsTotalCount += ITEMS_CHANGED_COUNT;
+			state.groupCreateStatus = DataStatus.FULFILLED;
+		});
+		builder.addCase(create.rejected, (state) => {
+			state.groupCreateStatus = DataStatus.REJECTED;
 		});
 	},
 	initialState,