+++ apps/frontend/src/pages/access-management/access-management.tsx
@@ -1,25 +1,30 @@
 import {
+	Button,
+	ConfirmationModal,
+	Modal,
 	PageLayout,
-	Table,
-	TablePagination,
 } from "~/libs/components/components.js";
 import { DataStatus } from "~/libs/enums/enums.js";
 import {
 	useAppDispatch,
 	useAppSelector,
+	useCallback,
 	useEffect,
+	useModal,
 	usePagination,
+	useState,
 } from "~/libs/hooks/hooks.js";
-import { actions as groupActions } from "~/modules/groups/groups.js";
+import {
+	actions as groupActions,
+	type GroupCreateRequestDto,
+} from "~/modules/groups/groups.js";
 import { actions as userActions } from "~/modules/users/users.js";
 
 import {
-	getGroupColumns,
-	getGroupRows,
-	getUserColumns,
-	getUserRows,
-} from "./libs/helpers/helpers.js";
-import { type GroupRow, type UserRow } from "./libs/types/types.js";
+	GroupCreateForm,
+	GroupsTable,
+	UsersTable,
+} from "./libs/components/components.js";
 import styles from "./styles.module.css";
 
 const AccessManagement = (): JSX.Element => {
@@ -32,6 +37,13 @@ const AccessManagement = (): JSX.Element => {
 	} = useAppSelector(({ users }) => users);
 
 	const {
+		dataStatus: groupsDataStatus,
+		groupDeleteStatus,
+		groups,
+		groupsTotalCount,
+	} = useAppSelector(({ groups }) => groups);
+
+	const {
 		onPageChange: onUserPageChange,
 		onPageSizeChange: onUserPageSizeChange,
 		page: userPage,
@@ -40,12 +52,6 @@ const AccessManagement = (): JSX.Element => {
 		queryParameterPrefix: "user",
 		totalItemsCount: usersTotalCount,
 	});
-
-	const {
-		dataStatus: groupsDataStatus,
-		groups,
-		groupsTotalCount,
-	} = useAppSelector(({ groups }) => groups);
 
 	const {
 		onPageChange: onGroupPageChange,
@@ -57,20 +63,84 @@ const AccessManagement = (): JSX.Element => {
 		totalItemsCount: groupsTotalCount,
 	});
 
+	const {
+		isOpened: isModalOpened,
+		onClose: onModalClose,
+		onOpen: onModalOpen,
+	} = useModal();
+
+	const {
+		isOpened: isDeleteModalOpen,
+		onClose: onDeleteModalClose,
+		onOpen: onDeleteModalOpen,
+	} = useModal();
+
+	const [selectedGroupId, setSelectedGroupId] = useState<null | number>(null);
+
+	const hasSelectedGroup = Boolean(selectedGroupId);
+
 	useEffect(() => {
 		void dispatch(
 			userActions.loadAll({ page: userPage, pageSize: userPageSize }),
 		);
+	}, [dispatch, userPage, userPageSize, onGroupPageChange]);
+
+	useEffect(() => {
 		void dispatch(
 			groupActions.loadAll({ page: groupPage, pageSize: groupPageSize }),
 		);
-	}, [dispatch, userPage, userPageSize, groupPage, groupPageSize]);
+	}, [dispatch, groupPage, groupPageSize, onGroupPageChange]);
 
-	const userColumns = getUserColumns();
-	const userData: UserRow[] = getUserRows(users);
+	useEffect(() => {
+		if (groupDeleteStatus === DataStatus.FULFILLED) {
+			const maxGroupPages = Math.ceil(groupsTotalCount / groupPageSize);
+			const maxUserPages = Math.ceil(usersTotalCount / userPageSize);
+			const validGroupPage = Math.min(groupPage, maxGroupPages);
+			const validUserPage = Math.min(userPage, maxUserPages);
+			onGroupPageChange(validGroupPage);
+			onUserPageChange(validUserPage);
+		}
+	}, [
+		groupDeleteStatus,
+		groupPage,
+		groupsTotalCount,
+		groupPageSize,
+		onGroupPageChange,
+		userPage,
+		usersTotalCount,
+		userPageSize,
+		onUserPageChange,
+	]);
 
-	const groupColumns = getGroupColumns();
-	const groupData: GroupRow[] = getGroupRows(groups);
+	const handleGroupCreateSubmit = useCallback(
+		(payload: GroupCreateRequestDto): void => {
+			void dispatch(
+				groupActions.create({
+					payload,
+					query: { page: userPage, pageSize: userPageSize },
+				}),
+			);
+			onModalClose();
+		},
+		[dispatch, onModalClose, userPage, userPageSize],
+	);
+
+	const handleGroupDelete = useCallback((): void => {
+		if (hasSelectedGroup) {
+			void dispatch(groupActions.deleteById({ id: selectedGroupId as number }));
+
+			setSelectedGroupId(null);
+			onDeleteModalClose();
+		}
+	}, [hasSelectedGroup, dispatch, selectedGroupId, onDeleteModalClose]);
+
+	const onDelete = useCallback(
+		(id: number): void => {
+			setSelectedGroupId(id);
+			onDeleteModalOpen();
+		},
+		[onDeleteModalOpen],
+	);
 
 	const isLoading = [usersDataStatus, groupsDataStatus].some(
 		(status) => status === DataStatus.IDLE || status === DataStatus.PENDING,
@@ -80,31 +150,48 @@ const AccessManagement = (): JSX.Element => {
 		<PageLayout isLoading={isLoading}>
 			<h1 className={styles["title"]}>Access Management</h1>
 			<section>
-				<h2 className={styles["section-title"]}>Users</h2>
-				<div className={styles["users-table"]}>
-					<Table<UserRow> columns={userColumns} data={userData} />
-					<TablePagination
-						onPageChange={onUserPageChange}
-						onPageSizeChange={onUserPageSizeChange}
-						page={userPage}
-						pageSize={userPageSize}
-						totalItemsCount={usersTotalCount}
-					/>
+				<div className={styles["section-header"]}>
+					<h2 className={styles["section-title"]}>Users</h2>
 				</div>
+				<UsersTable
+					onPageChange={onUserPageChange}
+					onPageSizeChange={onUserPageSizeChange}
+					page={userPage}
+					pageSize={userPageSize}
+					totalItemsCount={usersTotalCount}
+					users={users}
+				/>
 			</section>
 			<section>
-				<h2 className={styles["section-title"]}>Groups</h2>
-				<div className={styles["groups-table"]}>
-					<Table<GroupRow> columns={groupColumns} data={groupData} />
-					<TablePagination
-						onPageChange={onGroupPageChange}
-						onPageSizeChange={onGroupPageSizeChange}
-						page={groupPage}
-						pageSize={groupPageSize}
-						totalItemsCount={groupsTotalCount}
-					/>
+				<div className={styles["section-header"]}>
+					<h2 className={styles["section-title"]}>Groups</h2>
+					<Button label="Create New" onClick={onModalOpen} />
 				</div>
+				<GroupsTable
+					groups={groups}
+					onDelete={onDelete}
+					onPageChange={onGroupPageChange}
+					onPageSizeChange={onGroupPageSizeChange}
+					page={groupPage}
+					pageSize={groupPageSize}
+					totalItemsCount={groupsTotalCount}
+				/>
 			</section>
+			<Modal
+				isOpened={isModalOpened}
+				onClose={onModalClose}
+				title="Create new group"
+			>
+				<GroupCreateForm onSubmit={handleGroupCreateSubmit} />
+			</Modal>
+			{hasSelectedGroup && (
+				<ConfirmationModal
+					content="The group will be deleted. This action cannot be undone. Do you want to continue?"
+					isOpened={isDeleteModalOpen}
+					onClose={onDeleteModalClose}
+					onConfirm={handleGroupDelete}
+				/>
+			)}
 		</PageLayout>
 	);
 };