+++ packages/shared/src/libs/enums/api-path.enum.ts
@@ -2,6 +2,7 @@ const APIPath = {
 	ACTIVITY_LOGS: "/activity-logs",
 	AUTH: "/auth",
 	GROUPS: "/groups",
+	PERMISSIONS: "/permissions",
 	PROJECT_GROUPS: "/project-groups",
 	PROJECTS: "/projects",
 	USERS: "/users",