+++ apps/backend/CHANGELOG.md
@@ -1,5 +1,34 @@
 # Changelog
 
+## [1.12.1](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/backend-v1.12.0...backend-v1.12.1) (2024-09-10)
+
+
+### Bug Fixes
+
+* resctict user manage group gf-254 ([#256](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/256)) ([73c7473](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/73c74736ed11930b58260a95fdbea260a4ff40a6))
+
+## [1.12.0](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/backend-v1.11.0...backend-v1.12.0) (2024-09-10)
+
+
+### Features
+
+* add admin group gf-227 ([#235](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/235)) ([1114411](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/1114411c7ac28c715b13faaa27a790f7395758b5))
+* save search value to query parameters gf-200 ([#218](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/218)) ([8d5cb71](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/8d5cb7135a078181a191f653091cf0217cd61023))
+
+## [1.11.0](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/backend-v1.10.0...backend-v1.11.0) (2024-09-10)
+
+
+### Features
+
+* add an ablity to create a group gf-79 ([#176](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/176)) ([ad36683](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/ad366835d2fce03680d0d1208fde6ba3a56f5e57))
+
+## [1.10.0](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/backend-v1.9.0...backend-v1.10.0) (2024-09-09)
+
+
+### Features
+
+* restrict users manage groups gf-84 ([#175](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/175)) ([064e3dc](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/064e3dce0ecb217f380c313f4844c1b8b336b52a))
+
 ## [1.9.0](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/backend-v1.8.1...backend-v1.9.0) (2024-09-06)